refactor: 모듈 명 변경
